Rename trace_utils module to judge_trace_utils per ALKIS comment
- Renamed mlflow/genai/judges/trace_utils.py → judge_trace_utils.py
- Renamed tests/genai/judges/optimizers/test_trace_utils.py → test_judge_trace_utils.py
- Updated import statements in dspy_utils.py to use new module name
- Updated test file imports to use new module name
- Removed ALKIS comment now that task is completed
- Verified all tests pass with the new module structure
