+++ mlflow/genai/judges/optimizers/dspy_utils.py
@@ -5,7 +5,7 @@ from typing import Any
 
 from mlflow.entities.trace import Trace
 from mlflow.exceptions import MlflowException
-from mlflow.genai.judges.trace_utils import (
+from mlflow.genai.judges.judge_trace_utils import (
     extract_request_from_trace,
     extract_response_from_trace,
 )
